backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
pymysql==1.1.0
sympy==1.12
cryptography==42.0.5
